Implemented IComparable, added tests
