+++ ISOWeekDateTests/ISOWeekDateFormatTests.cs
@@ -7,24 +7,28 @@ namespace ISOWeekDate
 	public class ISOWeekDateFormatTests
 	{
 		[TestMethod()]
+		[TestCategory("Not Implemented")]
 		public void ShortCompactPatternTest()
 		{
 			throw new NotImplementedException();
 		}
 
 		[TestMethod()]
+		[TestCategory("Not Implemented")]
 		public void ShortExtendedPatternTest()
 		{
 			throw new NotImplementedException();
 		}
 
 		[TestMethod()]
+		[TestCategory("Not Implemented")]
 		public void FullCompactPatternTest()
 		{
 			throw new NotImplementedException();
 		}
 
 		[TestMethod()]
+		[TestCategory("Not Implemented")]
 		public void FullExtendedPatternTest()
 		{
 			throw new NotImplementedException();
